Crate ncollide3df32​

ncollide

ncollide is a n-dimensional collision detection library written with the rust programming language.

As its name suggests, it is generic wrt the dimension: it works with both 2-dimensional and 3-dimensional geometries. It might work with higher dimensions (never tried).

Compilation

You will need the last nightly build of the rust compiler and the official package manager: cargo.

Simply add the following to your Cargo.toml file:

[dependencies.ncollide3df32]
git = "https://github.com/sebcrozet/ncollide"

Features

dynamic bounding volume tree based broad phase

ball vs. ball collision detection,

plane vs. any convex object collision detection.

collision detection between arbitrary convex objects

compound geometries

ray-casting

time of impact computation for objects without rotational movement (compound vs. compound is not yet implemented)

And various traits for collision detectors and broad phase collision detection.

What is missing

Some common features are still missing:

heightmaps

Modules

bounding_volume
Bounding volumes.

broad
Broad phases.

data
Data structure utilities.

geom
Geometric primitives.

implicit
Definition of support functions.

math
Compilation flags dependent aliases for mathematical types.

narrow
Narrow phases.

parametric
Definition of parametric surfaces.

partitioning
Spatial partitioning tools.

procedural
Procedural mesh generation.

ray
Ray casting utilities.

utils
Miscelaneous, unsorted generic geometric utilities.

volumetric
Volume and inertia tensor computation.