Module ncollide3df32​::broad​

Broad phases.

Structs

BruteForceBoundingVolumeBroadPhase
Broad phase with quadratic complexity but sped up using loose bounding volumes.

BruteForceBroadPhase
Broad phase with quadratic complexity.

DBVTBroadPhase
Broad phase based on a Dynamic Bounding Volume Tree.

NoIdDispatcher
Dispatcher which disallows dispatches between identical pointers.

Traits

BoundingVolumeBroadPhase
Trait of broad phases working with bounding volume.

BroadPhase
Trait all broad phase must implement.

Dispatcher
Trait of dispatcher.

InterferencesBroadPhase
Thait of broad phases which check for pairwise interferences.

RayCastBroadPhase
Traits of broad phase able to run fast ray-cast queries.