Struct ncollide3df32​::broad​::BruteForceBroadPhase​

pub struct BruteForceBroadPhase<B, D, DV> {
    // some fields omitted
}

Broad phase with quadratic complexity.

It always returns false positives since it assumes that all object is in collision with all objects. Do not use this but for benchmarking the narrow phase.

Methods

impl<B: HasUid + Clone, D: Dispatcher<B, B, DV>, DV> BruteForceBroadPhase<B, D, DV>

fn new(dispatcher: D) -> BruteForceBroadPhase<B, D, DV>

Builds a new brute force broad phase.

fn pairs<'r>(&'r self) -> &'r HashMap<Pair<B>, DV, PairTWHash>

The pair manager of this broad phase.

fn add(&mut self, b: B)

Adds an element to this broad phase.

fn remove(&mut self, _: &B)

Removes an element from this broad phase.