Module ncollide3df32​::geom​

Geometric primitives.

Structs

AnnotatedMinkowskiSum
Same as the MinkowskiSum but with a support mapping which keeps track of the original supports points from the two wrapped geometries.

Ball
A Ball geometry.

BezierCurve
Procedural generator of non-rational Bézier curve.

BezierCurveEvaluationCache
Cache used to evaluate a bezier curve at a given parameter.

BezierSurface
Procedural generator of non-rational Bézier surfaces.

BezierSurfaceEvaluationCache
Cache used to evaluate a bezier surface at a given parameter.

Capsule
Implicit description of a capsule geometry with its principal axis aligned with the y axis.

Compound
A compound geometry with an aabb bounding volume.

CompoundData
Structure used to build a Compound geometry.

Cone
Implicit description of a cylinder geometry with its principal axis aligned with the y axis.

Convex
Set of point assumed to form a convex polytope.

Cuboid
Geometry of a box.

Cylinder
Implicit description of a cylinder geometry with its principal axis aligned with the y axis.

GeomWithMargin
Extends the wrapped geometry with its margin.

Mesh
Geometry commonly known as a 2d line strip or a 3d triangle mesh.

MinkowskiSum
Implicit representation of the Minkowski sum of two geometries.

Plane
Implicit description of a plane.

Reflection
Implicit representation of the reflection of a geometry.

Segment
A segment geometry.

Torus
A torus.

Triangle
A triangle geometry.

Traits

ConcaveGeom
Trait implemented by concave, composite geometries.

Geom
Trait (that should be) implemented by every geometry.

MeshElement
Trait implemented by elements usable on the Mesh.

Type Definitions

MeshPrimitive
The primitive geometry used by a Mesh.