Struct ncollide3df32​::geom​::Compound​

pub struct Compound {
    // some fields omitted
}

A compound geometry with an aabb bounding volume.

A compound geometry is a geometry composed of the union of several simpler geometry. This is the main way of creating a concave geometry from convex parts. Each parts can have its own delta transformation to shift or rotate it with regard to the other geometries.

Methods

impl Compound

fn new(data: CompoundData) -> Compound

Builds a new compound geometry.

impl Compound

fn geoms<'r>(&'r self) -> &'r [(Matrix, Arc<Box<Geom + Send + Sync>>)]

The geometries of this compound geometry.

fn bvt<'r>(&'r self) -> &'r BVT<uint, AABB>

The optimization structure used by this compound geometry.

fn bounding_volumes<'r>(&'r self) -> &'r [AABB]

The geometries bounding volumes.