Struct ncollide3df32​::geom​::CompoundData​

pub struct CompoundData {
    // some fields omitted
}

Structure used to build a Compound geometry.

This accumulates the geometries and their volumetric properties.

Methods

impl CompoundData

fn new() -> CompoundData

Creates a new CompoundData.

fn push_geom<S: Geom + Volumetric + Send + Sync>(&mut self, delta: Matrix, geom: S, density: Scalar)

Adds a new geometry; its mass properties are automatically computed.

fn push_geom_with_mass_properties<S: Geom + Send + Sync>(&mut self, delta: Matrix, geom: S, props: (Scalar, Scalar, Vect, AngularInertia))

Adds a new geometry with the given mass properties.

fn push_shared_geom_with_mass_properties(&mut self, delta: Matrix, geom: Arc<Box<Geom + Send + Sync>>, props: (Scalar, Scalar, Vect, AngularInertia))

Adds a new shared geometry with the given mass properties.

fn geoms<'r>(&'r self) -> &'r [(Matrix, Arc<Box<Geom + Send + Sync>>)]

The geometries stored by this CompoundData.

fn mass_properties_list<'r>(&'r self) -> &'r [(Scalar, Scalar, Vect, AngularInertia)]

The geometries stored by this CompoundData.