Struct ncollide3df32​::geom​::Segment​

pub struct Segment {
    // some fields omitted
}

A segment geometry.

Methods

impl Segment

fn new(a: Vect, b: Vect) -> Segment

Creates a new segment from two points.

The segment will have a default margin of 0.04.

fn new_with_margin(a: Vect, b: Vect, margin: Scalar) -> Segment

Creates a new segment from two points and a custom margin.

impl Segment

fn a<'a>(&'a self) -> &'a Vect

The first point of this segment.

fn b<'a>(&'a self) -> &'a Vect

The second point of this segment.

fn margin(&self) -> Scalar

The margin surrounding this segment.