Function ncollide3df32​::narrow​::algorithm​::gjk​::project_origin​

pub fn project_origin<S: Simplex<_V>, G: Implicit<_V, _M>, _V: FloatVec<Scalar>, _M>(m: &_M, geom: &G, simplex: &mut S) -> Option<_V>

Projects the origin on a geometry unsing the GJK algorithm.

Arguments:

geom - the geometry to project the origin on

simplex - the simplex to be used by the GJK algorithm. It must be already initialized with at least one point on the geometry boundary.