Struct ncollide3df32​::narrow​::surface_selector​::TangentConesSurfaceSelector​

pub struct TangentConesSurfaceSelector {
    // some fields omitted
}

A selector that tests the orthogonality condition using the surface tangent cone.

This is the criteria from Distance Extrema for Spline Models Using Tangent Cones, Ilijas Selimovic.

Methods

impl TangentConesSurfaceSelector

fn new(max_lmd: Scalar) -> TangentConesSurfaceSelector

Creates a new tangent-cone based surface detector.