Trait ncollide3df32​::narrow​::surface_selector​::SurfaceSelector​

pub trait SurfaceSelector<D> {
    fn set_max_lmd(&mut self, max_lmd: Scalar);
    fn is_flat(&mut self, surf: &BezierSurface, data: &D) -> bool;
    fn may_contain_a_closest_point(&mut self, pt: &Vect, b: &BezierSurface, data: &D) -> bool;
    fn create_test_data(&mut self, b: &BezierSurface) -> D;
}

Trait implemented by the heristics for surface selection during collision detection.

Required Methods

fn set_max_lmd(&mut self, max_lmd: Scalar)

Sets the maximum local minimal distance.

fn is_flat(&mut self, surf: &BezierSurface, data: &D) -> bool

Tells whether a surface is flat enough to run a numerical resolution algorithm.

fn may_contain_a_closest_point(&mut self, pt: &Vect, b: &BezierSurface, data: &D) -> bool

Tells whether a surface might contain a closest point or not.

fn create_test_data(&mut self, b: &BezierSurface) -> D

Allocates data used to test flatnass and closest point containment.