Struct ncollide3df32​::narrow​::surface_subdivision_tree​::SurfaceSubdivisionTree​

pub struct SurfaceSubdivisionTree<D> {
    // some fields omitted
}

A shareable binary tree with a pointer to its parent.

Methods

impl<D: Send + Sync> SurfaceSubdivisionTree<D>

fn new_orphan(b: BezierSurface, data: D, timestamp: uint) -> SurfaceSubdivisionTree<D>

Creates a new tree with no parent nor children.

fn surface<'a>(&'a self) -> &'a BezierSurface

The surface contained by this node.

fn data<'a>(&'a self) -> &'a D

Reference to the data contained by this node.

fn data_mut<'a>(&'a mut self) -> &'a mut D

Mutable reference to the data contained by this node.

fn timestamp(&self) -> uint

The timestamp of this tree node.

fn set_timestamp(&mut self, timestamp: uint)

Sets the timestamp of this tree node.

fn has_left_child(&self) -> bool

Whether or not this node has a left child.

fn has_right_child(&self) -> bool

Whether or not this node has a right child.

fn right_child(&self) -> Option<Arc<RWLock<SurfaceSubdivisionTree<D>>>>

A copy of this node right child.

fn left_child(&self) -> Option<Arc<RWLock<SurfaceSubdivisionTree<D>>>>

A copy of this node left child.

fn right_child_ref<'a>(&'a self) -> Option<&'a Arc<RWLock<SurfaceSubdivisionTree<D>>>>

A reference to this node right child.

fn left_child_ref<'a>(&'a self) -> Option<&'a Arc<RWLock<SurfaceSubdivisionTree<D>>>>

A reference to this node left child.

fn set_right_child(&mut self, child: SurfaceSubdivisionTree<D>)

Sets the right child of this node.

fn set_left_child(&mut self, child: SurfaceSubdivisionTree<D>)

Sets the left child of this node.

fn is_right_child(&self, child: &Arc<RWLock<SurfaceSubdivisionTree<D>>>) -> bool

Returns true if child is the right child of this node.

fn is_left_child(&self, child: &Arc<RWLock<SurfaceSubdivisionTree<D>>>) -> bool

Returns true if child is the left child of this node.

fn remove_right_child(&mut self)

Removes the right child of this node.

fn remove_left_child(&mut self)

Removes the left child of this node.