Struct ncollide3df32​::narrow​::surface_subdivision_tree​::SurfaceSubdivisionTreeRef​

pub struct SurfaceSubdivisionTreeRef<D> {
    // some fields omitted
}

A referenece to an element of the subdivision cache.

Each time an element is added to the cache, one of thoses references are created. The element will be kept in cache as long as at least one of those references exists.

Methods

impl<D> SurfaceSubdivisionTreeRef<D>

fn is_the_subdivision_tree_of(&self, b: &BezierSurface) -> bool

Tests if this references the subdivision tree of the bézier surface b.