Struct ncollide3df32​::partitioning​::BVT​

pub struct BVT<B, BV> {
    // some fields omitted
}

A Boundig Volume Tree.

Methods

impl<B, BV> BVT<B, BV>

fn new_with_partitioner(leaves: Vec<(B, BV)>, partitioner: |uint, Vec<(B, BV)>| -> (BV, BinaryPartition<B, BV>)) -> BVT<B, BV>

Builds a bounding volume tree using an user-defined construction function.

fn visit<Vis: BVTVisitor<B, BV>>(&self, visitor: &mut Vis)

Visit this tree using… a visitor!

fn visit_mut<Vis: BVTVisitor<B, BV>>(&mut self, visitor: &mut Vis)

Visit this tree using… a visitor! Visitor arguments are mutable.

fn visit_bvtt<Vis: BVTTVisitor<B, BV>>(&self, other: &BVT<B, BV>, visitor: &mut Vis)

Visits the bounding volume traversal tree implicitely formed with other.

fn root_bounding_volume<'r>(&'r self) -> Option<&'r BV>

Reference to the bounding volume of the tree root.

fn depth(&self) -> uint

Computes the depth of this tree.

impl<B, BV: Translation<Vect> + BoundingVolume + Clone> BVT<B, BV>

fn new_kdtree(leaves: Vec<(B, BV)>) -> BVT<B, BV>

Creates a new kdtree.

impl<B, BV: RayCast> BVT<B, BV>

fn cast_ray<'a, T>(&'a self, ray: &Ray, cast_fn: &mut |&B, &Ray| -> Option<(Scalar, T)>) -> Option<(Scalar, T, &'a B)>

Computes the closest intersection between the objects stored on this tree and a given ray.