Trait ncollide3df32​::partitioning​::BVTVisitor​

pub trait BVTVisitor<B, BV> {
    fn visit_internal(&mut self, &BV) -> bool;
    fn visit_leaf(&mut self, &B, &BV);

    fn visit_internal_mut(&mut self, bv: &mut BV) -> bool { ... }
    fn visit_leaf_mut(&mut self, b: &mut B, bv: &mut BV) { ... }
}

Visitor of Bounding Volume Trees.

Required Methods

fn visit_internal(&mut self, &BV) -> bool

Visits an internal node. Returns true if the internal node children have to be visited too.

fn visit_leaf(&mut self, &B, &BV)

Visits a leaf.

Provided Methods

fn visit_internal_mut(&mut self, bv: &mut BV) -> bool

Visits an internal node. Returns true if the internal node children have to be visited too.

fn visit_leaf_mut(&mut self, b: &mut B, bv: &mut BV)

Visits a leaf.