Module ncollide3df32​::ray​

Ray casting utilities.

Structs

Ray
A Ray.

RayIntersection
Structure containing the result of a successful ray cast.

Traits

RayCast
Traits of objects which can be tested for intersection with a ray.

Functions

ball_toi_with_ray
Computes the time of impact of a ray on a ball.

implicit_toi_and_normal_with_ray
Cast a ray on a geometry using the GJK algorithm.

plane_toi_with_ray
Computes the toi of a ray with a plane described by its center and normal.

triangle_ray_intersection
Computes the intersection between a triangle and a ray.