Function ncollide3df32​::utils​::triangulate​

pub fn triangulate<N: FloatMath + Cast<f64>, V: FloatVecExt<N> + Clone>(pts: &[V]) -> TriMesh<N, V>

Triangulates a set of point (sort of) lying on the same 2d plane.

If the points do not lie on the same 2d plane, strange things might happends (triangle might be attached together in an unnatural way). Though, if they are only slighly perturbated on the directions orthogonal to the plane, this should be fine.